// Anova.Evolve.Admin.Frontend/frontend/src/containers/AssetEditor/components/DigitalChannelForm/components/ObjectForm/types.ts
export interface Values {
  description?: string | null;
  rtuId?: string | null;
  rtuChannelId?: string | null;
  dataChannelTemplateId?: string | null;
  eventRuleGroupId?: number | null;
  state0Value?: number;
  state0Text?: string | null;
  state1Value?: number;
  state1Text?: string | null;
}